Add some notes
Struct is passed by value, class is passed by reference.
When passing a struct in the input of a function, there is an implicit
“let” in front of passed structs.
Structs include array, dictionary, int and so on.
And when copying a struct, Swift doesn’t really copy it. Swift is very
smart behind the scene. Swift pass it by a pointer or so called
reference and knows it is copied. Swift really copy it when something
change in the struct.